requirements.txt
Pillow
boto3
requests
numpy
scikit-image
